+++ op_crates/timers/Cargo.toml
@@ -1,0 +1,18 @@
+# Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
+
+[package]
+name = "deno_timers"
+version = "0.1.0"
+edition = "2018"
+description = "Timers API implementation for Deno"
+authors = ["the Deno authors"]
+license = "MIT"
+readme = "README.md"
+repository = "https://github.com/denoland/deno"
+
+[lib]
+path = "lib.rs"
+
+[dependencies]
+deno_core = { version = "0.84.0", path = "../../core" }
+tokio = { version = "1.4.0", features = ["full"] }
