+++ runtime/build.rs
@@ -13,8 +13,10 @@ fn create_snapshot(
   snapshot_path: &Path,
   files: Vec<PathBuf>,
 ) {
+  // Initialization order matters.
   deno_webidl::init(&mut js_runtime);
   deno_console::init(&mut js_runtime);
+  deno_timers::init(&mut js_runtime);
   deno_url::init(&mut js_runtime);
   deno_web::init(&mut js_runtime);
   deno_file::init(&mut js_runtime);
